update the coloring and font of the homepage
Google fonts were used to update the look of the homepage to be
more modern and pleasing to the eye. There was a slight deviation
from the colors and the functional design specification since
the homepage looks better with a darker shade of grey. These
slight deviations are reflected in the updated functional design
specification