finalize styling for homepage scroll animation
The opacity being set to zero is reserved for Javascript, in the
case Javascript isn't enabled the animated elements will remain
visible.

Consequently, the !important CSS rule is applied to the opacity of 1
set in CSS since it normally has lower precendence than the inline CSS
used when applying an opacity of 0 in Javascript.